ByteSpan seperated from comment handling
ByteSpan seperated from comment handling
