+++ sway-fmt-v2/src/utils/expr/abi_cast.rs
@@ -2,7 +2,7 @@ use crate::{
     fmt::*,
     utils::{
         bracket::Parenthesis,
-        comments::{ByteSpan, LeafSpans},
+        byte_span::{ByteSpan, LeafSpans},
     },
 };
 use std::fmt::Write;
