+++ sway-fmt-v2/src/utils/expr/assignable.rs
@@ -2,7 +2,7 @@ use crate::{
     fmt::*,
     utils::{
         bracket::SquareBracket,
-        comments::{ByteSpan, LeafSpans},
+        byte_span::{ByteSpan, LeafSpans},
     },
 };
 use std::fmt::Write;
