+++ sway-fmt-v2/src/utils/expr/collections.rs
@@ -2,7 +2,7 @@ use crate::{
     fmt::*,
     utils::{
         bracket::{Parenthesis, SquareBracket},
-        comments::{ByteSpan, LeafSpans},
+        byte_span::{ByteSpan, LeafSpans},
         shape::LineStyle,
     },
 };
