+++ sway-fmt-v2/src/utils/expr/conditional.rs
@@ -2,7 +2,7 @@ use crate::{
     fmt::*,
     utils::{
         bracket::CurlyBrace,
-        comments::{ByteSpan, LeafSpans},
+        byte_span::{ByteSpan, LeafSpans},
     },
 };
 use std::{fmt::Write, ops::ControlFlow};
