+++ sway-fmt-v2/src/utils/program_type.rs
@@ -1,5 +1,5 @@
 use crate::{
-    utils::comments::{ByteSpan, LeafSpans},
+    utils::byte_span::{ByteSpan, LeafSpans},
     FormatterError,
 };
 use std::fmt::Write;
